// web/src/api/generated/model/bankAccountDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface BankAccountDto {
  accountId: string;
  code: string;
  nameAr: string;
  nameEn: string;
  isActive: boolean;
  balance: number;
  reconciledBalance: number;
  unreconciled: number;
  /** @nullable */
  lastStatementDate: string | null;
  /** @nullable */
  lastStatementBalance: number | null;
}